+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Unsaved changes warning feature (Issue #324)
 - Fixes to filters and "All" option. It now never uses "*" as value, but all options in a {node1, node2, node3} expression (Issue #228, #359)
 - Fix for InfluxDB query generation with columns containing dots or dashes (Issue #369, #348) - Thanks to @jbripley
-
+- Improvement to series toggling, CTRL+MouseClick on series name will now hide all others (Issue #350)
 
 # 1.5.3 (2014-04-17)
 - Add support for async scripted dashboards (Issue #274)